+++ main/library_tree/properties.js
@@ -1,5 +1,5 @@
 ﻿'use strict';
-//05/04/26
+//07/04/26
 
 /* global $:readable */
 /* global globQuery:readable, globTags:readable */ // helpers\helpers_xxx_global.js
@@ -455,4 +455,4 @@ if (ppt.get('Tree List View')) {
 if (ppt.get('Image Type') > 5) { ppt.set('Image Type', 0); } // Regorxxx <- Custom TF art safety check ->
 ppt.set('Image Show Index Number', null);
 ppt.set('Image Show Index Year Auto', null);
-properties = undefined;
+properties = void(0);